requirements.txt
aiogram
aiohttp
aiosqlite
apscheduler
cachetools
openai
python-dotenv
